deps.ts standardization; normalizeFolderPath ; better path utils
