+++ flow-core/src/deps.ts
@@ -1,0 +1,5 @@
+export { normalize } from "https://deno.land/std@0.224.0/path/mod.ts";
+export { assertEquals, assertThrows, assertNotStrictEquals } from "https://deno.land/std@0.224.0/assert/mod.ts";
+export { basename, dirname, join, resolve } from "https://deno.land/std@0.224.0/path/mod.ts";
+export { existsSync } from "https://deno.land/std@0.224.0/fs/mod.ts";
+export { ensureDir } from "https://deno.land/std@0.224.0/fs/ensure_dir.ts";
