+++ flow-service/src/routes/meshes.ts
@@ -1,8 +1,9 @@
 import { OpenAPIHono, createRoute, z } from '@hono/zod-openapi';
 import { logger } from '../utils/logger.ts';
-import { getMetaFlowPath, getNextMetaDistPath, getCurrentMetaDistPath, MESH, getHandlePath, getAssetsPath } from '../../../flow-core/src/mesh-constants.ts';
-import { join, dirname } from 'jsr:@std/path';
-import { existsSync } from "https://deno.land/std@0.224.0/fs/mod.ts";
+import { getMetaFlowPath, getNextMetaDistPath, getCurrentMetaDistPath, getHandlePath, getAssetsPath } from '../../../flow-core/src/utils/mesh-path-utils.ts';
+import { normalizeFolderPath } from '../../../flow-core/src/utils/path-utils.ts';
+import { MESH } from '../../../flow-core/src/mesh-constants.ts';
+import { join, dirname, existsSync } from '../../../flow-core/src/deps.ts';
 import { ServiceConfigAccessor } from '../config/index.ts';
 import { composeMetadataContent } from '../services/metadata-composer.ts';
 import { initializeMeshRegistry } from '../utils/mesh-utils.ts';
@@ -120,7 +121,8 @@ export const createMeshesRoutes = (config: ServiceConfigAccessor): OpenAPIHono =
 
   meshes.openapi(registerMeshRoute, async (c) => {
     const { name, parentPath } = c.req.valid('json');
-    const path = join(parentPath, name);
+    const normalizedParentPath = normalizeFolderPath(parentPath);
+    const path = join(normalizedParentPath, name);
 
     if (meshRegistry[name]) {
       return c.json({
